renamed converters.py to stores.py
the main use of the classes is to store and retrieve the models. They also convert  the raw dictionary to a model, so it may be worth moving that logic if required in the future
